feature(7Rc0bcT4): added logic for create, solve, resolve, get and delete diagnostic tests

+++ src/main/java/pl/pjatk/learnease/entity/learn/Material.java
@@ -19,7 +19,7 @@ public class Material {
     @Id
     @Column(name = "material_id", unique = true)
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer materialId;
+    private Long materialId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "subject_id", nullable = false)
