+++ src/main/java/pl/pjatk/learnease/entity/user/Level.java
@@ -16,7 +16,7 @@ public class Level {
     @Id
     @Column(name = "level_id", unique = true)
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer levelId;
+    private Long levelId;
 
     @Column(name = "name", nullable = false)
     private String name;
